requirements.txt
# Only for testing pinned versions on the CI
appdirs==1.4.4
freezegun==1.1.0
httpx==0.18.1
pytablewriter[html]==0.60.0
pytest==6.2.4
pytest-cov==2.12.1
python-slugify==5.0.2
respx==0.17.0
termcolor==1.1.0
